Function numcodecs_bit_round::bit_round

pub fn bit_round<T: Float, S: Data<Elem = T>, D: Dimension>(
    data: ArrayBase<S, D>,
    keepbits: u8,
) -> Result<Array<T, D>, BitRoundCodecError>

Floating-point bit rounding.

Drops the specified number of bits from the floating point mantissa, leaving an array that is more amenable to compression. The number of bits to keep should be determined by information analysis of the data to be compressed.

The approach is based on the paper by Klöwer et al. 2021 (https://www.nature.com/articles/s43588-021-00156-2).

See https://github.com/milankl/BitInformation.jl for the the original implementation in Julia.

Errors

Errors with BitRoundCodecError::ExcessiveKeepBits if keepbits exceeds T::MANITSSA_BITS.