Struct numcodecs_bit_round::BitRoundCodec

pub struct BitRoundCodec {
    pub keepbits: u8,
}

Codec providing floating-point bit_rounding.

Fields
keepbits: u8

The number of bits of the mantissa to keep.

The valid range depends on the dtype of the input data.

If keepbits is equal to the bitlength of the dtypeโs mantissa, no transformation is performed.