numcodecs_bit_round

Trait Float

pub trait Float:
    Sized
    + Copy
    + Div<Self, Output = Self> {
    type Binary: Copy + Not<Output = Self::Binary> + Shr<u32, Output = Self::Binary> + Add<Self::Binary, Output = Self::Binary> + AddAssign<Self::Binary> + BitAnd<Self::Binary, Output = Self::Binary> + BitAndAssign<Self::Binary>;

    const MANITSSA_BITS: u32;
    const MANTISSA_MASK: Self::Binary;
    const BINARY_ONE: Self::Binary;
    const TY: AnyArrayDType;

    // Required methods
    fn to_binary(self) -> Self::Binary;
    fn from_binary(u: Self::Binary) -> Self;
    fn is_normal(self) -> bool;
    fn normal_log2_floor(self) -> i16;
    fn abs(self) -> Self;
    fn from_f64(x: f64) -> Self;
}

Floating point types.

Required Associated Constants

const MANITSSA_BITS: u32

Number of significant digits in base 2

const MANTISSA_MASK: Self::Binary

Binary mask to extract only the mantissa bits

const BINARY_ONE: Self::Binary

Binary 0x1

const TY: AnyArrayDType

Dtype of this type

Required Associated Types

type Binary: Copy + Not<Output = Self::Binary> + Shr<u32, Output = Self::Binary> + Add<Self::Binary, Output = Self::Binary> + AddAssign<Self::Binary> + BitAnd<Self::Binary, Output = Self::Binary> + BitAndAssign<Self::Binary>

Binary representation of this type

Required Methods

fn to_binary(self) -> Self::Binary

Bit-cast the floating point value to its binary representation

fn from_binary(u: Self::Binary) -> Self

Bit-cast the binary representation into a floating point value

fn is_normal(self) -> bool

Returns the floating point category of the number

fn normal_log2_floor(self) -> i16

Returns the floor of the base-2 logarithm as a signed integer

fn abs(self) -> Self

Computes the absolute value

fn from_f64(x: f64) -> Self

Convert from an f64 value

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".